+++ CHANGELOG.md
@@ -22,9 +22,12 @@ See for sample https://raw.githubusercontent.com/favoloso/conventional-changelog
 -->
 
 ## [0.10.0] - 2024-MM-DD
+### ✨ Feature
+- Display Favicons (#38)
+### 🛠 Improvements
+- Refactor media storage for scalability (#44)
 ### 🐛 Bug Fixes
 - Fix admin % lighthouse inQueue value (#42)
-- Refactor media storage for scalability (#44)
 
 ## [0.9.2] - 2024-05-20
 ### Improvements